requirements.txt
python-telegram-bot==21.3
psycopg2-binary==2.9.7
